requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
aiohttp==3.9.5
python-dotenv==1.0.1
jinja2==3.1.4
websockets==12.0
pydantic==2.7.0
